+++ service/database/uploadLogo.go
@@ -8,7 +8,7 @@ import (
 )
 
 func (db *appdbimpl) UploadLogo(p Photo, u User) (Photo, User, error) {
-	var photoId, profilePic uint64
+	var photoId, profilePic int
 	var userName, followers, banned, photos string
 
 	//search for the user
@@ -81,7 +81,7 @@ func (db *appdbimpl) UploadLogo(p Photo, u User) (Photo, User, error) {
 			return p, u, err
 		}
 
-		p.ID = uint64(lastInsertID)
+		p.ID = int(lastInsertID)
 
 		//We update the user
 		u.Name = userName
